+++ homework_scripts/3.sh
@@ -14,12 +14,19 @@ COMMON_ARGS=(
 )
 
 NPROC=2
-
-echo -e "Delete previous experiments\n"
-rm -rf outputs/fsdp-full-shard-noCPU outputs/fsdp-full-shard-CPU outputs/fsdp-full-shard-CPU-checkpointing \
-       outputs/fsdp-full-shard-noCPU.log outputs/fsdp-full-shard-CPU.log outputs/fsdp-full-shard-CPU-checkpointing.log
-echo -e "Previous experiments deleted\n"
-
+is_deleting_previous=0
+while [ $# -gt 0 ]; do
+    case "$1" in
+        -c|--clean) is_deleting_previous=1; shift ;;
+        *) echo "Unknown arg: $1"; exit 1 ;;
+    esac
+done
+if [ "$is_deleting_previous" -eq 1]; then
+    echo -e "Delete previous experiments\n"
+    rm -rf outputs/fsdp-full-shard-noCPU outputs/fsdp-full-shard-CPU outputs/fsdp-full-shard-CPU-checkpointing \
+         outputs/fsdp-full-shard-noCPU.log outputs/fsdp-full-shard-CPU.log outputs/fsdp-full-shard-CPU-checkpointing.log
+    echo -e "Previous experiments deleted\n"
+fi
 # 1. FULL SHARD (NO CPU)
 echo -e "Run experiment\tFULL SHARD NO CPU\n===============================\n"
 uv run torchrun --standalone --nproc_per_node=${NPROC} train_fsdp.py \
